refactor(diff): DiffSearchPanel 拆分搜索算法 / 着色 / 状态机
DiffSearchPanel.tsx 576 → ~190 行，退化为瘦渲染组件。新增 search/ 子目录：

- search/diff-search.ts：纯跨文件搜索逻辑（runSearch + findChangedIndices /
  expandToContext / stripLeadingIndent / findMatchSpan / countLines / loadContent）
  + 类型 LineMatch / FileResults + 常量，无 React 依赖
- search/colorize.ts：colorizeAll（Monaco editor.colorize 异步语法着色）
- search/useDiffSearch.ts：搜索状态机 hook（query / 大小写持久化 / 结果 /
  loading / 折叠态 + 去抖搜索 + 内容缓存 + 自动聚焦）

纯结构拆分，行为不变；DiffSearchPanel 导出名与路径保持不变（DiffView 引用零波及）。
